requirements.txt
tensorflow==2.20.0
numpy==2.2.6
Pillow==12.0.0
scikit-image==0.25.2
scipy==1.15.3